requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pycricbuzz>=0.1
duckduckgo-search>=4.0
googlesearch-python>=1.2.0
google-generativeai>=0.3.0
groq>=0.4.0
python-dotenv>=1.0.0
